fix(gateway): tighten pairing.approved token delivery (#418)
